+++ src/core/lombok/eclipse/handlers/HandleWithBy.java
@@ -252,7 +252,7 @@ public class HandleWithBy extends EclipseAnnotationHandler<WithBy> {
 		if (method.returnType == null) return null;
 		
 		Annotation[] deprecated = null, checkerFramework = null;
-		if (isFieldDeprecated(fieldNode)) deprecated = new Annotation[] { generateDeprecatedAnnotation(source) };
+		if (isFieldDeprecated(fieldNode)) deprecated = new Annotation[] { generateDeprecatedAnnotation(source, fieldNode) };
 		if (getCheckerFrameworkVersion(fieldNode).generateSideEffectFree()) checkerFramework = new Annotation[] { generateNamedAnnotation(source, CheckerFrameworkVersion.NAME__SIDE_EFFECT_FREE) };
 		
 		char[][] functionalInterfaceName = null;