+++ src/core/lombok/eclipse/handlers/singulars/EclipseGuavaSingularizer.java
@@ -130,7 +130,7 @@ abstract class EclipseGuavaSingularizer extends EclipseSingularizer {
 		md.statements = returnStatement != null ? new Statement[] {a, returnStatement} : new Statement[] {a};
 		md.returnType = returnType;
 		addCheckerFrameworkReturnsReceiver(md.returnType, data.getSource(), cfv);
-		md.annotations = generateSelfReturnAnnotations(deprecate, data.getSource());
+		md.annotations = generateSelfReturnAnnotations(deprecate, data, data.getSource());
 		
 		data.setGeneratedByRecursive(md);
 		if (returnStatement != null) createRelevantNonNullAnnotation(builderType, md);
@@ -177,7 +177,7 @@ abstract class EclipseGuavaSingularizer extends EclipseSingularizer {
 		addCheckerFrameworkReturnsReceiver(md.returnType, data.getSource(), cfv);
 		char[] prefixedSingularName = data.getSetterPrefix().length == 0 ? data.getSingularName() : HandlerUtil.buildAccessorName(builderType, new String(data.getSetterPrefix()), new String(data.getSingularName())).toCharArray();
 		md.selector = fluent ? prefixedSingularName : HandlerUtil.buildAccessorName(builderType, "add", new String(data.getSingularName())).toCharArray();
-		Annotation[] selfReturnAnnotations = generateSelfReturnAnnotations(deprecate, data.getSource());
+		Annotation[] selfReturnAnnotations = generateSelfReturnAnnotations(deprecate, data, data.getSource());
 		Annotation[] copyToSetterAnnotations = copyAnnotations(md, findCopyableToBuilderSingularSetterAnnotations(data.getAnnotation().up()));
 		md.annotations = concat(selfReturnAnnotations, copyToSetterAnnotations, Annotation.class);
 
@@ -218,7 +218,7 @@ abstract class EclipseGuavaSingularizer extends EclipseSingularizer {
 		addCheckerFrameworkReturnsReceiver(md.returnType, data.getSource(), cfv);
 		char[] prefixedSelector = data.getSetterPrefix().length == 0 ? data.getPluralName() : HandlerUtil.buildAccessorName(builderType, new String(data.getSetterPrefix()), new String(data.getPluralName())).toCharArray();
 		md.selector = fluent ? prefixedSelector : HandlerUtil.buildAccessorName(builderType, "addAll", new String(data.getPluralName())).toCharArray();
-		Annotation[] selfReturnAnnotations = generateSelfReturnAnnotations(deprecate, data.getSource());
+		Annotation[] selfReturnAnnotations = generateSelfReturnAnnotations(deprecate, data, data.getSource());
 		Annotation[] copyToSetterAnnotations = copyAnnotations(md, findCopyableToSetterAnnotations(data.getAnnotation().up(), true));
 		md.annotations = concat(selfReturnAnnotations, copyToSetterAnnotations, Annotation.class);