+++ src/core/lombok/javac/handlers/HandleWith.java
@@ -282,7 +282,7 @@ public class HandleWith extends JavacAnnotationHandler<With> {
 		CheckerFrameworkVersion checkerFramework = getCheckerFrameworkVersion(source);
 		if (checkerFramework.generateSideEffectFree()) annsOnMethod = annsOnMethod.prepend(maker.Annotation(genTypeRef(source, CheckerFrameworkVersion.NAME__SIDE_EFFECT_FREE), List.<JCExpression>nil()));
 		
-		if (isFieldDeprecated(field)) annsOnMethod = annsOnMethod.prepend(maker.Annotation(genJavaLangTypeRef(field, "Deprecated"), List.<JCExpression>nil()));
+		if (isFieldDeprecated(field)) annsOnMethod = annsOnMethod.prepend(generateDeprecatedAnnotation(field));
 		
 		if (makeAbstract) access |= Flags.ABSTRACT;
 		AnnotationValues<Accessors> accessors = JavacHandlerUtil.getAccessorsForField(field);