+++ src/core/lombok/javac/handlers/HandleWithBy.java
@@ -325,7 +325,7 @@ public class HandleWithBy extends JavacAnnotationHandler<WithBy> {
 		CheckerFrameworkVersion checkerFramework = getCheckerFrameworkVersion(source);
 		if (checkerFramework.generateSideEffectFree()) annsOnMethod = annsOnMethod.prepend(maker.Annotation(genTypeRef(source, CheckerFrameworkVersion.NAME__SIDE_EFFECT_FREE), List.<JCExpression>nil()));
 		
-		if (isFieldDeprecated(field)) annsOnMethod = annsOnMethod.prepend(maker.Annotation(genJavaLangTypeRef(field, "Deprecated"), List.<JCExpression>nil()));
+		if (isFieldDeprecated(field)) annsOnMethod = annsOnMethod.prepend(generateDeprecatedAnnotation(field));
 		
 		if (makeAbstract) access = access | Flags.ABSTRACT;
 		AnnotationValues<Accessors> accessors = JavacHandlerUtil.getAccessorsForField(field);